增加saga注解模式向前重试功能

+++ src/main/java/icu/funkye/easy/tx/config/annotation/GlobalTransaction.java
@@ -19,7 +19,7 @@ public @interface GlobalTransaction {
 
     EasyTxMode mode() default EasyTxMode.EASY ;
 
-    long timeout() default 5000;
+    int timeout() default 5000;
 
     /**
      * 是否重试
